backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
pandas==2.2.3
numpy==2.2.1
pyarrow==18.1.0
yfinance>=1.1.0
ta==0.11.0
pytz==2024.2
pyyaml==6.0.2
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
